move structures/classes that is no need to share and make new APIs for hiding C++ codes
